standalone/A7 Cross Site Scripting lab/requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1
crispy_bootstrap4
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.1.7
django-allauth==0.52.0
django-crispy-forms==2.0
django-heroku==0.3.1
gunicorn==23.0.0
idna==3.4
mccabe==0.6.1
oauthlib==3.2.2
Pillow==9.4.0
psycopg2==2.9.3
pycodestyle==2.7.0
pycparser==2.21
pyflakes==2.3.1
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.1
requests==2.28.2
requests-oauthlib==1.3.1
sqlparse==0.3.1
urllib3==1.26.9
Werkzeug==2.1.2
whitenoise==6.2.0
zipp==3.8.0
